add. ClipCompressCommand

+++ src/main/kotlin/com/helltar/twitchviewer_bot/commands/commands/ClipCommand.kt
@@ -3,15 +3,12 @@ package com.helltar.twitchviewer_bot.commands.commands
 import com.github.kotlintelegrambot.Bot
 import com.github.kotlintelegrambot.entities.Message
 import com.github.kotlintelegrambot.entities.TelegramFile
-import com.helltar.twitchviewer_bot.BotConfig
 import com.helltar.twitchviewer_bot.Strings
-import com.helltar.twitchviewer_bot.commands.TwitchCommand
 import com.helltar.twitchviewer_bot.twitch.Twitch
 import com.helltar.twitchviewer_bot.utils.Utils
-import org.slf4j.LoggerFactory
 import java.io.File
 
-class ClipCommand(bot: Bot, message: Message, args: List<String> = listOf()) : TwitchCommand(bot, message, args) {
+class ClipCommand(bot: Bot, message: Message, args: List<String> = listOf()) : ClipCompressCommand(bot, message, args) {
 
     override fun run() {
         if (args.isEmpty())
@@ -20,14 +17,16 @@ class ClipCommand(bot: Bot, message: Message, args: List<String> = listOf()) : T
             else
                 sendMessage(localizedString(Strings.clip_command_info))
         else
-            if (checkIsChannelNameValid())
-                getClipsFromAll(listOf(args[0]))
+            if (checkIsChannelNameValid()) {
+                val compress = args.size > 1 && args[1].startsWith(".")
+                getClipsFromAll(listOf(args[0]), compress)
+            }
     }
 
-    fun getClipsFromAll(userLogins: List<String>) {
+    fun getClipsFromAll(userLogins: List<String>, compress: Boolean = false) {
         twitch.getOnlineList(userLogins)?.let {
             if (it.isNotEmpty())
-                sendClip(it)
+                sendClip(it, compress)
             else {
                 val text = if (userLogins.size > 1)
                     localizedString(Strings.empty_online_list)
@@ -36,14 +35,11 @@ class ClipCommand(bot: Bot, message: Message, args: List<String> = listOf()) : T
 
                 sendMessage(text)
             }
-
-            return
         }
-
-        sendMessage(localizedString(Strings.twitch_exception))
+            ?: sendMessage(localizedString(Strings.twitch_exception))
     }
 
-    private fun sendClip(twitchStreamsData: List<Twitch.StreamsData>) {
+    private fun sendClip(twitchStreamsData: List<Twitch.StreamsData>, compress: Boolean) {
         twitchStreamsData.forEach {
             val tempMessageId = sendMessage(String.format(localizedString(Strings.start_get_clip), it.username))
             val filename = twitch.getShortClip(it.login)
@@ -54,23 +50,22 @@ class ClipCommand(bot: Bot, message: Message, args: List<String> = listOf()) : T
                 return
             }
 
-            val gameName = if (it.gameName.isNotEmpty())
-                ", #${Utils.replaceTitleTag(it.gameName)}"
-            else ""
+            if (!compress) {
+                val gameName = if (it.gameName.isNotEmpty())
+                    ", #${Utils.replaceTitleTag(it.gameName)}"
+                else ""
 
-            bot.sendVideo(
-                chatId,
-                TelegramFile.ByFile(File(filename)),
-                15, 1920, 1080, // -c copy
-                caption = "#${it.username}$gameName",
-                replyToMessageId = replyToMessageId, allowSendingWithoutReply = true
-            )
+                bot.sendVideo(
+                    chatId,
+                    TelegramFile.ByFile(File(filename)),
+                    15, 1920, 1080, // -c copy
+                    caption = "#${it.username}$gameName",
+                    replyToMessageId = replyToMessageId, allowSendingWithoutReply = true
+                )
 
-            try {
                 File(filename).delete()
-            } catch (e: Exception) {
-                LoggerFactory.getLogger(javaClass)
-            }
+            } else
+                compressAndSendVideo(filename)
         }
     }
 }
